re-work hostdb reliability code
Switched the code to a float from uint64 so that precision could remain strong
even when the values are tiny. This means we can apply decay even when the
values are tiny.

Test updated to reflect the clamps and newer decay methods and strengths.

+++ modules/renter.go
@@ -98,10 +98,10 @@ type HostDBEntry struct {
 	HistoricUptime   time.Duration `json:"historicuptime"`
 	ScanHistory      HostDBScans   `json:"scanhistory"`
 
-	HistoricFailedInteractions     uint64 `json:"historicfailedinteractions"`
-	HistoricSuccessfulInteractions uint64 `json:"historicsuccessfulinteractions"`
-	RecentFailedInteractions       uint64 `json:"recentfailedinteractions"`
-	RecentSuccessfulInteractions   uint64 `json:"recentsuccessfulinteractions"`
+	HistoricFailedInteractions     float64 `json:"historicfailedinteractions"`
+	HistoricSuccessfulInteractions float64 `json:"historicsuccessfulinteractions"`
+	RecentFailedInteractions       float64 `json:"recentfailedinteractions"`
+	RecentSuccessfulInteractions   float64 `json:"recentsuccessfulinteractions"`
 
 	LastHistoricUpdate types.BlockHeight
